Revert "cdh: move kms crate to cdh module"
We use the KMS crate outside of the guest-components. Specifically, we
use some of this functionality in Trustee as part of resource backends
that use KMSes.

We could potentially adjust Trustee to import the CDH and use the KMSes
from there, but Trustee doesn't need anything else from the CDH.

It probably does make sense to have the KMS logic in its own crate,
hence revert the change that brough it into the CDH.

This reverts commit 3ab81292768eea451531bbc0ab7f38a3fe94aa7c.
