+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
     "attestation-agent/deps/sev",
     "attestation-agent/coco_keyprovider",
     "confidential-data-hub/hub",
+    "confidential-data-hub/kms",
     "image-rs",
     "ocicrypt-rs",
 ]
