+++ confidential-data-hub/hub/src/image/error.rs
@@ -5,8 +5,6 @@
 
 use thiserror::Error;
 
-use crate::kms;
-
 pub type Result<T> = std::result::Result<T, Error>;
 
 #[derive(Error, Debug)]
