+++ confidential-data-hub/hub/src/lib.rs
@@ -17,6 +17,5 @@ pub mod config;
 pub use config::*;
 
 pub mod image;
-pub mod kms;
 pub mod secret;
 pub mod storage;
