+++ confidential-data-hub/kms/Cargo.toml
@@ -1,0 +1,59 @@
+[package]
+name = "kms"
+version = "0.1.0"
+authors = ["The Confidential Container Authors"]
+publish = false
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+attestation-agent = { path = "../../attestation-agent/attestation-agent", default-features = false }
+base64.workspace = true
+bincode = { workspace = true, optional = true }
+chrono = { workspace = true, optional = true }
+const_format.workspace = true
+crypto = { path = "../../attestation-agent/deps/crypto", optional = true }
+ehsm_client = {git = "https://github.com/intel/ehsm", rev = "3454cac66b968a593c3edc43410c0b52416bbd3e", optional = true }
+hex = { workspace = true, optional = true }
+kbs_protocol = { path = "../../attestation-agent/kbs_protocol", default-features = false, features = [
+    "passport",
+    "aa_ttrpc",
+    "openssl",
+], optional = true }
+log.workspace = true
+p12 = { version = "0.6.3", optional = true }
+prost = { workspace = true, optional = true }
+rand = { workspace = true, optional = true }
+reqwest = { workspace = true, optional = true }
+resource_uri = { path = "../../attestation-agent/deps/resource_uri" }
+ring = "0.17"
+sha2 = { workspace = true, optional = true }
+serde.workspace = true
+serde_json.workspace = true
+sev = { path = "../../attestation-agent/deps/sev", optional = true }
+strum.workspace = true
+thiserror.workspace = true
+tokio = { workspace = true, features = ["fs"] }
+toml.workspace = true
+tonic = { workspace = true, optional = true }
+url = { workspace = true, optional = true }
+uuid = { workspace = true, features = ["serde", "v4"], optional = true }
+yasna = { version = "0.5.2", optional = true }
+zeroize = { workspace = true, optional = true }
+
+[dev-dependencies]
+rstest.workspace = true
+tokio = { workspace = true, features = ["rt", "macros" ] }
+
+[build-dependencies]
+anyhow.workspace = true
+tonic-build.workspace = true
+
+[features]
+default = ["aliyun", "kbs"]
+
+aliyun = ["chrono", "hex", "p12", "prost", "rand", "reqwest/rustls-tls", "sha2", "tonic", "url", "yasna"]
+kbs = ["kbs_protocol"]
+ehsm = ["ehsm_client"]
+sev = ["bincode", "crypto", "dep:sev", "prost", "tonic", "uuid", "zeroize"]
