+++ confidential-data-hub/kms/build.rs
@@ -1,0 +1,20 @@
+#![allow(missing_docs)]
+
+// extern crate tonic_build;
+
+use anyhow::*;
+
+fn main() -> Result<()> {
+    #[cfg(feature = "aliyun")]
+    tonic_build::compile_protos(
+        "./src/plugins/aliyun/client/client_key_client/protobuf/dkms_api.proto",
+    )?;
+
+    #[cfg(feature = "sev")]
+    tonic_build::configure()
+        .build_server(true)
+        .out_dir("./src/plugins/kbs/sev")
+        .compile_protos(&["./src/plugins/kbs/sev/protos/getsecret.proto"], &[""])?;
+
+    Ok(())
+}
