+++ confidential-data-hub/kms/src/error.rs
