+++ confidential-data-hub/kms/src/plugins/aliyun/client/sts_token_client/credential.rs
