+++ confidential-data-hub/kms/src/plugins/aliyun/client/sts_token_client/mod.rs
@@ -20,7 +20,7 @@ use serde::Deserialize;
 use serde_json::Value;
 use tokio::fs;
 
-use crate::kms::{
+use crate::{
     error::{Error, Result},
     plugins::aliyun::annotations::AliSecretAnnotations,
     Annotations, ProviderSettings,
