+++ confidential-data-hub/kms/src/plugins/ehsm/annotations.rs
