+++ confidential-data-hub/kms/src/plugins/ehsm/example_credential/credential.4eb1____.json
