+++ confidential-data-hub/kms/src/plugins/kbs/cc_kbc.rs
@@ -13,7 +13,7 @@ use kbs_protocol::{
 };
 use log::{info, warn};
 
-use super::{Error, Result};
+use crate::{Error, Result};
 
 use super::Kbc;
