+++ confidential-data-hub/kms/src/plugins/kbs/sev/protos/getsecret.proto
