+++ confidential-data-hub/kms/src/plugins/mod.rs
@@ -7,7 +7,7 @@ use std::str::FromStr;
 
 use strum::{AsRefStr, EnumString};
 
-use super::{Decrypter, Error, Getter, ProviderSettings, Result};
+use crate::{Decrypter, Error, Getter, ProviderSettings, Result};
 
 const _IN_GUEST_DEFAULT_KEY_PATH: &str = "/run/confidential-containers/cdh/kms-credential";
